refactor: 컴포넌트 테스트 클래스 패키지 이동 및 이름 변경
- 컴포넌트 테스트 클래스를 `woowacourse.payments.ui` 패키지에서
`woowacourse.payments.ui.component` 패키지로 이동했습니다.

- 또한, `CardExpirationDateTextFieldTest` 클래스에서 오류 메시지 "유효하지 않은 만료일 입니다."를 "유효하지 않은 만료일입니다."로 수정했습니다.
