+++ app/src/androidTest/java/woowacourse/payments/ui/component/CardPasswordTextFieldTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.component
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -16,7 +16,6 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.component.CardPasswordTextField
 
 class CardPasswordTextFieldTest {
     @get:Rule
